# openapi_specs/openapi3.yml
openapi: 3.0.1
info:
  title: VAmPI
  description: OpenAPI v3 specs for VAmPI
  version: '0.1'
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: models.user_model.User.decode_auth_token
paths:
...
  /books/v1:
    get:
      tags:
        - books
      summary: Retrieves all books
      description: Retrieves all books
      operationId: api_views.books.get_all_books
      responses:
        '200':
          description: See all books
          content:
            application/json:
              schema:
                type: object
                properties:
                  Books:
                    type: array
                    items:
                      type: object
                      properties:
                        book_title:
                          type: string
                        user:
                          type: string
              example:
                Books:
                  - book_title: 'bookTitle77'
                    user: 'name1'
                  - book_title: 'bookTitle85'
                    user: 'name2'
                  - book_title: 'bookTitle47'
                    user: 'admin'
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - books
      summary: Add new book
      description: Add new book
      operationId: api_views.books.add_new_book
      requestBody:
        description: >-
          Add new book with title and secret content only available to the user
          who added it.
        content:
          application/json:
            schema:
              type: object
              properties:
                book_title:
                  type: string
                  example: 'book99'
                secret:
                  type: string
                  example: 'pass1secret'
        required: true
      responses:
        '200':
          description: Successfully added a book
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Book has been added.'
                  status:
                    type: string
                    enum: ['success', 'fail']
                    example: 'success'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: ['fail']
                    example: 'fail'
                  message:
                    type: string
                    example: 'Book Already exists!'
        '401':
          description: User not authorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: ['fail']
                    example: 'fail'
                  message:
                    type: string
                    example: 'Invalid token. Please log in again.'
  /books/v1/{book_title}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - books
      summary: Retrieves book by title along with secret
      description: >-
        Retrieves book by title along with secret. Only the owner may retrieve
        it
      operationId: api_views.books.get_by_title
      parameters:
        - name: book_title
          in: path
          description: retrieve book data
          required: true
          schema:
            type: string
            example: 'bookTitle77'
      responses:
        '200':
          description: Successfully retrieve book info
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    book_title:
                      type: string
                      example: 'bookTitle77'
                    owner:
                      type: string
                      example: 'name1'
                    secret:
                      type: string
                      example: 'secret for bookTitle77'
        '401':
          description: User not authorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: ['fail']
                    example: 'fail'
                  message:
                    type: string
                    example: 'Invalid token. Please log in again.'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: ['fail']
                    example: 'fail'
                  message:
                    type: string
                    example: 'Book not found!'
  /books/v1/{book_title}/cover:
    post:
      security:
        - bearerAuth: []
      tags:
        - books
      summary: Set a book cover from a remote URL
      description: >-
        Fetches the cover image from the given URL server-side and stores its
        metadata against the book.
      operationId: api_views.books.set_cover
      parameters:
        - name: book_title
          in: path
          description: book to set the cover for
          required: true
          schema:
            type: string
            example: 'bookTitle77'
      requestBody:
        description: Remote URL to fetch the cover image from
        content:
          application/json:
            schema:
              type: object
              properties:
                cover_url:
                  type: string
                  example: 'https://example.test/covers/bookTitle77.png'
        required: true
      responses:
        '200':
          description: Successfully fetched the cover
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: 'success'
                  content_type:
                    type: string
                    example: 'image/png'
                  size:
                    type: number
                    example: 10240
        '400':
          description: Invalid request
          content: {}
        '401':
          description: User not authorized
          content: {}
  /books/v1/{book_title}/export:
    get:
      security:
        - bearerAuth: []
      tags:
        - books
      summary: Export the book's stored report file
      description: Reads the named export file for the book from the exports directory.
      operationId: api_views.books.export_book
      parameters:
        - name: book_title
          in: path
          required: true
          schema:
            type: string
            example: 'bookTitle77'
        - name: file
          in: query
          description: export file name, defaults to {book_title}.txt
          required: false
          schema:
            type: string
            example: 'bookTitle77.txt'
      responses:
        '200':
          description: Successfully exported the book report
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: 'success'
                  content:
                    type: string
        '401':
          description: User not authorized
          content: {}
        '404':
          description: Export file not found
          content: {}
...